# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Email Test Server API",
    "description": "API for Email Test Server",
    "version": "1.0.0"
  },
  "paths": {
...
    "/email": {
      "get": {
        "tags": ["Email"],
        "summary": "Get all emails",
        "operationId": "getEmails",
        "responses": {
          "200": {
            "description": "List of emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Email"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new email",
        "operationId": "createEmail",
        "requestBody": {
          "description": "Email data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Email"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Email created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/email/{email_id}": {
      "get": {
        "tags": ["Email"],
        "summary": "Get an email by ID",
        "operationId": "getEmailById",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "description": "Email ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "404": {
            "description": "Email not found"
          }
        }
      },
      "put": {
        "tags": ["Email"],
        "summary": "Update an existing email",
        "operationId": "updateEmail",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "description": "Email ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated email data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Email"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Email not found"
          }
        }
      },
      "delete": {
        "tags": ["Email"],
        "summary": "Delete an email by ID",
        "operationId": "deleteEmail",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "description": "Email ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email deleted successfully"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/email/ph/create": {
      "tags": ["Email"],
      "post": {
        "summary": "Generate phishing email logs",
        "operationId": "generatePhishingLogs",
        "requestBody": {
          "description": "Phishing log data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhishingLog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phishing logs generated"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/email/ph/send/{training_id}": {
      "tags": ["Email"],
      "post": {
        "summary": "Send phishing emails for a training",
        "operationId": "sendPhishingEmails",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "description": "Training ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Employee list for phishing emails",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employees": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phishing emails sent"
          },
          "400": {
            "description": "Employee list is missing"
          }
        }
      }
    },
...
    "/role": {
      "get": {
        "tags": ["Role"],
        "summary": "Get all roles",
        "operationId": "getAllRoles",
        "responses": {
          "200": {
            "description": "List of roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Role"],
        "summary": "Create a new role",
        "operationId": "createRole",
        "requestBody": {
          "description": "Role data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/role/{id}": {
      "get": {
        "summary": "Get role by ID",
        "operationId": "getRoleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing role",
        "operationId": "updateRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated role data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Role not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a role by ID",
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully"
          },
          "404": {
            "description": "Role not found"
          }
        }
      }
    },
...
  }
}